+++ packages/core/src/__tests__/name-generator.test.ts
@@ -1,235 +1,256 @@
-import type { TestScenario } from "@webmangler/testing";
-import type { CharSet } from "../characters";
+import type { TestScenarios } from "@webmangler/testing";
+import type { CharSet } from "@webmangler/types";
 
 import { expect } from "chai";
 
-import {
-  ALL_LOWERCASE_CHARS,
-  ALL_NUMBER_CHARS,
-  ALL_UPPERCASE_CHARS,
-} from "../characters";
 import NameGenerator from "../name-generator.class";
 
 interface TestCase {
-  charSet?: CharSet;
+  charSet: CharSet;
   reserved?: string[];
   samples: {
-    inc: number,
-    expected: string
+    inc: number;
+    expected: string;
   }[];
 }
 
+const ALL_LOWERCASE_CHARS = [
+  "a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l", "m", "n", "o",
+  "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z",
+];
+
+const ALL_NUMBER_CHARS = [
+  "0", "1", "2", "3", "4", "5", "6", "7", "8", "9",
+];
+
+const ALL_UPPERCASE_CHARS = [
+  "A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O",
+  "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z",
+];
+
 suite("NameGenerator", function() {
-  const defaultCharsetArray = Array.from(ALL_LOWERCASE_CHARS);
-
-  const allLowercaseCharsArray = Array.from(ALL_LOWERCASE_CHARS);
-  const allNumberCharsArray = Array.from(ALL_NUMBER_CHARS);
-  const allUppercaseCharsArray = Array.from(ALL_UPPERCASE_CHARS);
-
-  const reservedScenarios: TestScenario<TestCase>[] = [
+  const reservedScenarios: TestScenarios<TestCase[]> = [
     {
-      name: "no reserved",
-      cases: [
+      testName: "no reserved",
+      getScenario: () => [
         {
+          charSet: ALL_LOWERCASE_CHARS as CharSet,
           samples: [
-            ...defaultCharsetArray.map((char) => {
+            ...ALL_LOWERCASE_CHARS.map((char) => {
               return { inc: 0, expected: char };
             }),
             { inc: 0, expected: "aa" },
             { inc: 0, expected: "ab" },
-            { inc: defaultCharsetArray.length - 3, expected: "az" },
+            { inc: ALL_LOWERCASE_CHARS.length - 3, expected: "az" },
             { inc: 0, expected: "ba" },
           ],
         },
         {
+          charSet: ALL_LOWERCASE_CHARS as CharSet,
           samples: [
-            { inc: defaultCharsetArray.length**2, expected: "za" },
-            { inc: defaultCharsetArray.length - 2, expected: "zz" },
+            { inc: ALL_LOWERCASE_CHARS.length**2, expected: "za" },
+            { inc: ALL_LOWERCASE_CHARS.length - 2, expected: "zz" },
             { inc: 0, expected: "aaa" },
             { inc: 0, expected: "aab" },
           ],
         },
         {
-          charSet: [...allLowercaseCharsArray, ...allUppercaseCharsArray],
+          charSet: [
+            ...ALL_LOWERCASE_CHARS,
+            ...ALL_UPPERCASE_CHARS,
+          ] as CharSet,
           samples: [
-            ...allLowercaseCharsArray.map((char) => {
+            ...ALL_LOWERCASE_CHARS.map((char) => {
               return { inc: 0, expected: char };
             }),
-            ...allUppercaseCharsArray.map((char) => {
+            ...ALL_UPPERCASE_CHARS.map((char) => {
               return { inc: 0, expected: char };
             }),
             { inc: 0, expected: "aa" },
-            { inc: allLowercaseCharsArray.length - 1, expected: "aA" },
+            { inc: ALL_LOWERCASE_CHARS.length - 1, expected: "aA" },
             { inc: 0, expected: "aB" },
-            { inc: allUppercaseCharsArray.length - 2, expected: "ba" },
-            { inc: allLowercaseCharsArray.length - 1, expected: "bA" },
+            { inc: ALL_UPPERCASE_CHARS.length - 2, expected: "ba" },
+            { inc: ALL_LOWERCASE_CHARS.length - 1, expected: "bA" },
           ],
         },
         {
-          charSet: [...allLowercaseCharsArray, ...allNumberCharsArray],
+          charSet: [
+            ...ALL_LOWERCASE_CHARS,
+            ...ALL_NUMBER_CHARS,
+          ] as CharSet,
           samples: [
-            ...allLowercaseCharsArray.map((char) => {
+            ...ALL_LOWERCASE_CHARS.map((char) => {
               return { inc: 0, expected: char };
             }),
-            ...allNumberCharsArray.map((char) => {
+            ...ALL_NUMBER_CHARS.map((char) => {
               return { inc: 0, expected: char };
             }),
             { inc: 0, expected: "aa" },
-            { inc: allLowercaseCharsArray.length - 1, expected: "a0" },
+            { inc: ALL_LOWERCASE_CHARS.length - 1, expected: "a0" },
             { inc: 1, expected: "a2" },
-            { inc: allNumberCharsArray.length - 3, expected: "ba" },
-            { inc: allLowercaseCharsArray.length - 1, expected: "b0" },
+            { inc: ALL_NUMBER_CHARS.length - 3, expected: "ba" },
+            { inc: ALL_LOWERCASE_CHARS.length - 1, expected: "b0" },
           ],
         },
       ],
     },
     {
-      name: "reserved strings",
-      cases: [
+      testName: "reserved strings",
+      getScenario: () => [
         {
+          charSet: ALL_LOWERCASE_CHARS as CharSet,
           reserved: ["a"],
           samples: [
             { inc: 0, expected: "b" },
-            { inc: defaultCharsetArray.length - 2, expected: "aa" },
+            { inc: ALL_LOWERCASE_CHARS.length - 2, expected: "aa" },
           ],
         },
         {
+          charSet: ALL_LOWERCASE_CHARS as CharSet,
           reserved: ["a", "b", "ab"],
           samples: [
             { inc: 0, expected: "c" },
-            { inc: defaultCharsetArray.length - 3, expected: "aa" },
+            { inc: ALL_LOWERCASE_CHARS.length - 3, expected: "aa" },
             { inc: 0, expected: "ac" },
           ],
         },
         {
-          charSet: allLowercaseCharsArray,
+          charSet: ALL_LOWERCASE_CHARS as CharSet,
           reserved: ["a"],
           samples: [
             { inc: 0, expected: "b" },
-            { inc: allLowercaseCharsArray.length - 2, expected: "aa" },
+            { inc: ALL_LOWERCASE_CHARS.length - 2, expected: "aa" },
           ],
         },
         {
-          charSet: allUppercaseCharsArray,
+          charSet: ALL_UPPERCASE_CHARS as CharSet,
           reserved: ["B"],
           samples: [
             { inc: 1, expected: "C" },
-            { inc: allUppercaseCharsArray.length - 2, expected: "AB" },
+            { inc: ALL_UPPERCASE_CHARS.length - 2, expected: "AB" },
           ],
         },
         {
-          charSet: allNumberCharsArray,
+          charSet: ALL_NUMBER_CHARS as CharSet,
           reserved: ["3", "1", "4"],
           samples: [
             { inc: 2, expected: "5" },
-            { inc: allNumberCharsArray.length - 5, expected: "01" },
+            { inc: ALL_NUMBER_CHARS.length - 5, expected: "01" },
           ],
         },
       ],
     },
     {
-      name: "reserved patterns",
-      cases: [
+      testName: "reserved patterns",
+      getScenario: () => [
         {
+          charSet: ALL_LOWERCASE_CHARS as CharSet,
           reserved: ["a.*"],
           samples: [
             { inc: 0, expected: "b" },
-            { inc: defaultCharsetArray.length - 2, expected: "ba" },
+            { inc: ALL_LOWERCASE_CHARS.length - 2, expected: "ba" },
           ],
         },
         {
+          charSet: ALL_LOWERCASE_CHARS as CharSet,
           reserved: [".*b"],
           samples: [
             { inc: 0, expected: "a" },
             { inc: 0, expected: "c" },
-            { inc: defaultCharsetArray.length - 3, expected: "aa" },
+            { inc: ALL_LOWERCASE_CHARS.length - 3, expected: "aa" },
             { inc: 0, expected: "ac" },
           ],
         },
         {
+          charSet: ALL_LOWERCASE_CHARS as CharSet,
           reserved: ["a.", ".c"],
           samples: [
             { inc: 0, expected: "a" },
             { inc: 0, expected: "b" },
             { inc: 0, expected: "c" },
-            { inc: defaultCharsetArray.length - 3, expected: "ba" },
+            { inc: ALL_LOWERCASE_CHARS.length - 3, expected: "ba" },
             { inc: 1, expected: "bd" },
           ],
         },
         {
-          charSet: allNumberCharsArray,
+          charSet: ALL_NUMBER_CHARS as CharSet,
           reserved: ["0.*"],
           samples: [
             { inc: 0, expected: "1" },
-            { inc: allNumberCharsArray.length - 2, expected: "10" },
+            { inc: ALL_NUMBER_CHARS.length - 2, expected: "10" },
           ],
         },
         {
-          charSet: allUppercaseCharsArray,
+          charSet: ALL_UPPERCASE_CHARS as CharSet,
           reserved: [".*B.*"],
           samples: [
             { inc: 1, expected: "C" },
-            { inc: allUppercaseCharsArray.length - 2, expected: "AC" },
-            { inc: allUppercaseCharsArray.length - 3, expected: "CA" },
+            { inc: ALL_UPPERCASE_CHARS.length - 2, expected: "AC" },
+            { inc: ALL_UPPERCASE_CHARS.length - 3, expected: "CA" },
           ],
         },
         {
-          charSet: [...allLowercaseCharsArray, ...allUppercaseCharsArray],
+          charSet: [
+            ...ALL_LOWERCASE_CHARS,
+            ...ALL_UPPERCASE_CHARS,
+          ] as CharSet,
           reserved: ["[b-d]", "[D-G]"],
           samples: [
             { inc: 1, expected: "e" },
-            { inc: allLowercaseCharsArray.length - 3, expected: "C" },
+            { inc: ALL_LOWERCASE_CHARS.length - 3, expected: "C" },
             { inc: 0, expected: "H" },
           ],
         },
       ],
     },
     {
-      name: "reserved strings & patterns",
-      cases: [
+      testName: "reserved strings & patterns",
+      getScenario: () => [
         {
+          charSet: ALL_LOWERCASE_CHARS as CharSet,
           reserved: ["a.*", "c"],
           samples: [
             { inc: 0, expected: "b" },
             { inc: 0, expected: "d" },
-            { inc: defaultCharsetArray.length - 4, expected: "ba" },
+            { inc: ALL_LOWERCASE_CHARS.length - 4, expected: "ba" },
             { inc: 1, expected: "bc" },
           ],
         },
         {
+          charSet: ALL_LOWERCASE_CHARS as CharSet,
           reserved: ["c", ".+b"],
           samples: [
             { inc: 0, expected: "a" },
             { inc: 0, expected: "b" },
-            { inc: defaultCharsetArray.length - 3, expected: "aa" },
+            { inc: ALL_LOWERCASE_CHARS.length - 3, expected: "aa" },
             { inc: 0, expected: "ac" },
-            { inc: defaultCharsetArray.length - 2, expected: "bc" },
+            { inc: ALL_LOWERCASE_CHARS.length - 2, expected: "bc" },
           ],
         },
         {
-          charSet: allNumberCharsArray,
+          charSet: ALL_NUMBER_CHARS as CharSet,
           reserved: ["0.*", "2"],
           samples: [
             { inc: 0, expected: "1" },
             { inc: 0, expected: "3" },
-            { inc: allNumberCharsArray.length - 4, expected: "10" },
+            { inc: ALL_NUMBER_CHARS.length - 4, expected: "10" },
           ],
         },
         {
-          charSet: allUppercaseCharsArray,
+          charSet: ALL_UPPERCASE_CHARS as CharSet,
           reserved: ["E", ".*B.*"],
           samples: [
             { inc: 1, expected: "C" },
             { inc: 1, expected: "F" },
-            { inc: allUppercaseCharsArray.length - 5, expected: "AC" },
-            { inc: allUppercaseCharsArray.length - 3, expected: "CA" },
+            { inc: ALL_UPPERCASE_CHARS.length - 5, expected: "AC" },
+            { inc: ALL_UPPERCASE_CHARS.length - 3, expected: "CA" },
           ],
         },
       ],
     },
     {
-      name: "duplicate characters",
-      cases: [
+      testName: "duplicate characters",
+      getScenario: () => [
         {
           charSet: ["a", "b", "c", "a"],
           samples: [
@@ -255,16 +276,15 @@ suite("NameGenerator", function() {
     },
   ];
 
-  for (const { name, cases } of reservedScenarios) {
-    test(name, function() {
-      for (const testCase of cases) {
-        const { charSet: _charSet, reserved, samples } = testCase;
-        const charSet = _charSet || defaultCharsetArray;
+  for (const { getScenario, testName } of reservedScenarios) {
+    test(testName, function() {
+      for (const testCase of getScenario()) {
+        const { charSet, reserved, samples } = testCase;
         expect(charSet).to.have.length.above(3);
 
         const g = new NameGenerator({
           reservedNames: reserved,
-          charSet: _charSet,
+          charSet,
         });
 
         for (const { inc, expected } of samples) {