+++ packages/core/src/embeds.ts
@@ -1,11 +1,10 @@
-import type { CharSet } from "./characters";
 import type {
+  CharSet,
   WebManglerEmbed,
   WebManglerFile,
   WebManglerLanguagePlugin,
-} from "./types";
+} from "@webmangler/types";
 
-import { ALL_LETTER_CHARS, ALL_NUMBER_CHARS } from "./characters";
 import NameGenerator from "./name-generator.class";
 
 type EmbedsMap = Map<WebManglerFile, Iterable<IdentifiableWebManglerEmbed>>;
@@ -28,8 +27,11 @@ export interface IdentifiableWebManglerEmbed extends WebManglerEmbed {
  * The {@link CharSet} used to generate unique identifiers for embed locations.
  */
 const idCharSet: CharSet = [
-  ...ALL_LETTER_CHARS,
-  ...ALL_NUMBER_CHARS,
+  "a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l", "m", "n", "o",
+  "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z", "A", "B", "C", "D",
+  "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S",
+  "T", "U", "V", "W", "X", "Y", "Z", "0", "1", "2", "3", "4", "5", "6", "7",
+  "8", "9",
 ];
 
 /**