+++ packages/language-html/src/embeds/__tests__/common/types.d.ts
@@ -1,0 +1,10 @@
+import type { WebManglerEmbed, WebManglerFile } from "@webmangler/types";
+
+interface TestCase {
+  file: WebManglerFile;
+  expected: WebManglerEmbed[];
+}
+
+export type {
+  TestCase,
+};