+++ packages/language-html/src/embeds/__tests__/css.test.ts
@@ -1,9 +1,0 @@
-import { expect } from "chai";
-
-import embeddedCssFinders from "../css";
-
-suite("HTML CSS Embeds", function() {
-  test("exported values", function() {
-    expect(embeddedCssFinders).to.have.at.length.greaterThan(0);
-  });
-});