+++ packages/language-html/src/embeds/__tests__/types.d.ts
@@ -1,6 +1,0 @@
-import type { WebManglerEmbed, WebManglerFile } from "@webmangler/types";
-
-export type TestCase = {
-  file: WebManglerFile;
-  expected: WebManglerEmbed[];
-}