+++ packages/language-html/src/embeds/css/__tests__/unit/common.test.ts
@@ -1,0 +1,9 @@
+import { expect } from "chai";
+
+import * as common from "../../common";
+
+suite("HTML CSS Embeds - Common", function() {
+  test("embed type", function() {
+    expect(common.EMBED_TYPE).to.equal("css");
+  });
+});