+++ packages/language-js/CHANGELOG.md
@@ -10,6 +10,10 @@ Versioning].
 
 - _No changes yet_
 
+## [0.1.27] - 2022-01-29
+
+- Bump minimum `@webmangler/language-utils` version to `^0.1.26`. ([#250])
+
 ## [0.1.26] - 2022-01-15
 
 - Fix query selectors not being mangled due to escaped quotes. ([#237])
@@ -32,5 +36,6 @@ Versioning].
 [#139]: https://github.com/ericcornelissen/webmangler/pull/139
 [#224]: https://github.com/ericcornelissen/webmangler/pull/224
 [#237]: https://github.com/ericcornelissen/webmangler/pull/237
+[#250]: https://github.com/ericcornelissen/webmangler/pull/250
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/ "Keep a CHANGELOG"
 [semantic versioning]: https://semver.org/spec/v2.0.0.html "Semantic versioning"